+++ src/redis/mod.rs
@@ -14,15 +14,17 @@ impl RedisClient {
     /// Connect to redis
     pub fn connect(url: &str) -> RedisResult<Self> {
         let client = Client::open(url)?;
+        debug!("connected to {}", url);
         Ok(Self { client })
     }
 
     /// Set key
     pub async fn set<V>(&mut self, key: &str, value: V) -> RedisResult<()>
     where
-        V: ToRedisArgs + Send + Sync,
+        V: ToRedisArgs + Send + Sync + std::fmt::Debug,
     {
         let mut connection = self.client.get_async_connection().await?;
+        debug!("SET {} to {:?}", key, value);
         connection.set(key, value).await
     }
 
@@ -32,27 +34,31 @@ impl RedisClient {
         V: FromRedisValue,
     {
         let mut connection = self.client.get_async_connection().await?;
+        debug!("GET {}", key);
         connection.get(key).await
     }
 
     /// Get all keys matching `pattern`
     pub async fn keys(&mut self, pattern: &str) -> RedisResult<Vec<String>> {
         let mut connection = self.client.get_async_connection().await?;
+        debug!("KEYS {}", pattern);
         connection.keys(pattern).await
     }
 
     /// Delete key
     pub async fn delete(&mut self, key: &str) -> RedisResult<()> {
         let mut connection = self.client.get_async_connection().await?;
+        debug!("DEL {}", key);
         connection.del(key).await
     }
 
     /// Push `value` to list at `key`
     pub async fn list_push<V>(&mut self, key: &str, value: V) -> RedisResult<()>
     where
-        V: ToRedisArgs + Send + Sync,
+        V: ToRedisArgs + Send + Sync + std::fmt::Debug,
     {
         let mut connection = self.client.get_async_connection().await?;
+        debug!("RPUSH {} {:?}", key, value);
         connection.rpush(key, value).await
     }
 
@@ -62,12 +68,14 @@ impl RedisClient {
         V: FromRedisValue,
     {
         let mut connection = self.client.get_async_connection().await?;
+        debug!("LINDEX {} {}", key, index);
         connection.lindex(key, index).await
     }
 
     /// Get list length
     pub async fn list_len(&mut self, key: &str) -> RedisResult<usize> {
         let mut connection = self.client.get_async_connection().await?;
+        debug!("LLEN {}", key);
         connection.llen(key).await
     }
 }